api/requirements.txt
# Vercel Serverless Function Requirements
# Python dependencies for API functions

httpx==0.26.0
requests==2.31.0
huggingface-hub>=0.20.0
Pillow==10.2.0
python-dotenv==1.0.0
pydantic==2.5.0
